add AutoOrValue, remove DoubleOrAuto, IntOrAuto
